feat: deploy contracts to testnet, upgrade to soroban-sdk v26, lazy DB init for Vercel
- Circle contract: CBQ5AFJXUHHPTYZ2CREDNTS4E5NMJJHUKQBKITGY4FURHB4KCBGT3KR7
- Reputation contract: CDRBHNJZVNBKW2VO3FUAH6A6UBWMBTMURNS5LHUOL5GUJNCC2I5M5A7Y
- Upgrade soroban-sdk 21.0.0 -> 26.0.0 for wasm32v1-none target compatibility
- Remove unused String import from circle contract
- Lazy PrismaClient init so backend starts without DATABASE_URL
- Health check now reports contract IDs and DB status
- Frontend: https://frontend-coral-nine-24.vercel.app
- Backend: https://backend-nine-eta-58.vercel.app

+++ packages/contracts/circle/src/lib.rs
@@ -1,5 +1,5 @@
 #![no_std]
-use soroban_sdk::{contract, contractimpl, contracttype, Address, Env, Symbol, Vec, String};
+use soroban_sdk::{contract, contractimpl, contracttype, Address, Env, Symbol, Vec};
 
 #[contracttype]
 #[derive(Clone)]
@@ -37,7 +37,7 @@ pub struct CycleState {
 }
 
 #[contracttype]
-#[derive(Clone, Debug, Copy)]
+#[derive(Clone, Debug)]
 pub enum DataKey {
     Circle,
     Member(Address),
@@ -241,7 +241,7 @@ impl CircleContract {
         // Advance to next cycle if not at end
         if config.current_cycle + 1 < config.total_members {
             config.current_cycle += 1;
-            let next_recipient = config.payout_order.get(config.current_cycle as usize)
+            let next_recipient = config.payout_order.get(config.current_cycle)
                 .expect("Invalid payout order index");
             let next_deadline = env.ledger().timestamp() + (config.cycle_length_days as u64 * 86400);
